% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundR2-package.R
\docType{package}
\name{roundR2-package}
\alias{roundR2}
\alias{roundR2-package}
\title{roundR2: Round Numeric Values to the Nearest Integer}
\description{
In R the default choice for rounding follows the IEC 60559 standard where a 5 is expected to be rounded to the even digit. This package offers alternative functionality to round to the closest integer.
}
\keyword{internal}
